@@ -3,31 +3,53 @@ description: A new Flutter project.
 version: 6.1.2+31
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  charts_flutter: ^0.6.0
-  dart_notification_center: ^1.0.0
+  #charts_flutter: ^0.6.0
+  charts_flutter: ^0.11.0
+  #dart_notification_center: ^1.0.0
+  dart_notification_center: ^1.0.0+1
   dropdown_banner: ^1.4.0
-  http: ^0.12.0+1
-  intl: ^0.15.8
-  auto_size_text: ^1.1.2
-  time_machine: ^0.9.9
-  flutter_slidable: ^0.5.0
-  firebase_analytics: ^4.0.2
-  firebase_core: ^0.4.0+8
-  firebase_messaging: ^5.1.2
-  flutter_apns: ^1.0.5
-  flutter_keychain: ^1.0.0
-  package_info: ^0.4.0+4
-  path_provider: ^1.1.2
-  share: ^0.6.1+1
-  shared_preferences: ^0.5.3+1
-  url_launcher: ^5.4.1
-  app_review: ^1.0.0
-  webview_flutter: ^0.3.11
-  camera: ^0.5.5+1
-  image: ^2.1.4
+  #http: ^0.12.0+1
+  http: ^0.13.3
+  #image_picker: ^0.6.3+1
+  image_picker: ^0.8.2
+  #intl: ^0.15.8
+  intl: ^0.17.0
+  #auto_size_text: ^1.1.2
+  auto_size_text: ^2.1.0
+  #time_machine: ^0.9.9
+  time_machine: ^0.9.16
+  #flutter_slidable: ^0.5.0
+  flutter_slidable: ^0.6.0
+  #firebase_analytics: ^4.0.2
+  firebase_analytics: ^8.2.0
+  #firebase_core: ^0.4.0+8
+  firebase_core: ^1.4.0
+  #firebase_messaging: ^5.1.2
+  firebase_messaging: ^10.0.4
+  #flutter_apns: ^1.0.5
+  flutter_apns: ^1.5.1
+  #flutter_keychain: ^1.0.0
+  flutter_keychain: ^2.0.1
+  package_info: ^2.0.2
+  path_provider: ^2.0.2
+  share: ^2.0.4
+  shared_preferences: ^2.0.6
+  url_launcher: ^6.0.9
+  app_review: ^2.1.1+1
+  webview_flutter: ^2.0.10
+  camera: ^0.9.1
+  image: ^3.0.2
+  flutter_multi_formatter: ^2.3.5
+  fluttertoast: ^7.1.1
+  rxdart: ^0.27.2
+  stripe_payment: ^1.0.11
+  screen_loader: ^2.0.1
+  pay: ^1.0.6
+
+
   flutter:
     sdk: flutter
 
@@ -71,5 +93,6 @@ flutter:
     - image_assets/tutorial_assets/
     - image_assets/share_assets/
     - image_assets/todolist_gif_assets/
+    - assets/
     - packages/time_machine/data/cultures/cultures.bin
     - packages/time_machine/data/tzdb/tzdb.bin
