@@ -2,20 +2,23 @@ part of 'constants.dart';
 
 class SKButton extends StatelessWidget {
   final String buttonText;
-  final EdgeInsets margins;
+  final EdgeInsets? margins;
   final double width;
+  final bool isDark;
   final ContextCallback callback;
 
   SKButton(
-      {Key key,
-      String buttonText,
-      EdgeInsets margin,
-      double width,
-      ContextCallback callback})
-      : buttonText = buttonText,
+      {Key? key,
+      String? buttonText,
+      EdgeInsets? margin,
+      double? width,
+      ContextCallback? callback,
+      bool isDark = false})
+      : buttonText = buttonText!,
         margins = margin,
-        width = width,
-        callback = callback,
+        width = width!,
+        callback = callback!,
+        isDark = isDark,
         super(key: key);
 
   Widget build(BuildContext context) => Container(
@@ -24,7 +27,7 @@ class SKButton extends StatelessWidget {
         width: width,
         child: SizedBox.expand(
           child: RaisedButton(
-            color: Colors.white,
+            color: isDark ? SKColors.dark_gray : Colors.white,
             shape:
                 RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
             child: Text(
@@ -32,7 +35,7 @@ class SKButton extends StatelessWidget {
               style: TextStyle(
                   fontSize: 15,
                   fontWeight: FontWeight.bold,
-                  color: SKColors.skoller_blue),
+                  color: isDark ? Colors.white : SKColors.skoller_blue1),
             ),
             onPressed: () {
               callback(context);
@@ -46,16 +49,16 @@ class SKNavBar extends StatelessWidget {
   final bool leftIsPop;
 
   final String title;
-  final Widget titleOption;
+  final Widget? titleOption;
 
-  final Color titleColor;
+  final Color? titleColor;
 
-  final Widget leftBtn;
-  final Widget rightBtn;
+  final Widget? leftBtn;
+  final Widget? rightBtn;
 
-  final VoidCallback callbackRight;
-  final VoidCallback callbackLeft;
-  final VoidCallback callbackTitle;
+  final VoidCallback? callbackRight;
+  final VoidCallback? callbackLeft;
+  final VoidCallback? callbackTitle;
 
   SKNavBar(
     this.title, {
@@ -73,7 +76,7 @@ class SKNavBar extends StatelessWidget {
     final titleWidget = GestureDetector(
       behavior: HitTestBehavior.opaque,
       onTapUp: (details) {
-        if (callbackTitle != null) callbackTitle();
+        if (callbackTitle != null) callbackTitle!();
       },
       child: Text(
         title,
@@ -102,7 +105,7 @@ class SKNavBar extends StatelessWidget {
             behavior: HitTestBehavior.opaque,
             onTapUp: (details) {
               if (callbackLeft != null)
-                callbackLeft();
+                callbackLeft!();
               else if (leftIsPop) Navigator.pop(context);
             },
             child: Container(
@@ -116,11 +119,11 @@ class SKNavBar extends StatelessWidget {
               ? titleWidget
               : Row(
                   mainAxisSize: MainAxisSize.min,
-                  children: <Widget>[titleWidget, titleOption]),
+                  children: <Widget>[titleWidget, titleOption!]),
           GestureDetector(
             behavior: HitTestBehavior.opaque,
             onTapUp: (details) {
-              if (callbackRight != null) callbackRight();
+              if (callbackRight != null) callbackRight!();
             },
             child: Container(
               padding: EdgeInsets.only(right: 4),
@@ -139,23 +142,23 @@ class SKNavView extends StatelessWidget {
   final bool isPop;
 
   final String title;
-  final Widget titleOption;
+  final Widget? titleOption;
 
-  final Widget rightBtn;
-  final Widget leftBtn;
+  final Widget? rightBtn;
+  final Widget? leftBtn;
 
-  final VoidCallback callbackRight;
-  final VoidCallback callbackLeft;
-  final VoidCallback callbackTitle;
+  final VoidCallback? callbackRight;
+  final VoidCallback? callbackLeft;
+  final VoidCallback? callbackTitle;
 
   final List<Widget> children;
 
-  final Color titleColor;
-  final Color backgroundColor;
+  final Color? titleColor;
+  final Color? backgroundColor;
 
   SKNavView({
-    @required this.children,
-    @required this.title,
+    required this.children,
+    required this.title,
     this.titleColor,
     this.rightBtn,
     this.isPop = true,
@@ -271,13 +274,13 @@ class SKCalendarPicker extends StatefulWidget {
   @override
   State createState() => _SKCalendarPickerState();
 
-  static Future<DateTime> presentDateSelector({
-    @required String title,
-    @required String subtitle,
-    @required BuildContext context,
-    @required DateTime startDate,
-    DateCallback onSelect,
-    DateContextCallback onSave,
+  static Future<dynamic> presentDateSelector({
+    required String title,
+    required String subtitle,
+    required BuildContext context,
+    required DateTime startDate,
+    DateCallback? onSelect,
+    DateContextCallback? onSave,
     bool showNoDate = false,
   }) async {
     DateTime selectedDate = startDate;
@@ -346,7 +349,7 @@ class SKCalendarPicker extends StatefulWidget {
                           EdgeInsets.symmetric(vertical: 8, horizontal: 12),
                       child: GestureDetector(
                         onTapUp: (_) {
-                          onSelect(null);
+                          onSelect!(DateTime.now());
                           Navigator.pop(context);
                         },
                         child: Text(
@@ -382,9 +385,9 @@ class SKCalendarPicker extends StatefulWidget {
                           onTapUp: (details) async {
                             // Wait to save if not used in a data-passing context
                             if (isSave)
-                              await onSave(selectedDate, context);
+                              await onSave!(selectedDate, context);
                             else
-                              onSelect(selectedDate);
+                              onSelect!(selectedDate);
 
                             Navigator.pop(context);
                           },
@@ -419,7 +422,7 @@ class _SKCalendarPickerState extends State<SKCalendarPicker> {
   List<DateTime> children = [];
   int curIndex = 1;
 
-  DateTime selectedDate;
+  late DateTime selectedDate;
 
   @override
   void initState() {
@@ -553,11 +556,11 @@ class _SKCalendarPickerState extends State<SKCalendarPicker> {
 class _SKCalendarBody extends StatefulWidget {
   final DateTime firstOfMonth;
   final DateTime startDate;
-  final DateCallback onDateSelected;
-  final DateFetch getSelected;
+  final DateCallback? onDateSelected;
+  final DateFetch? getSelected;
 
-  _SKCalendarBody({DateTime month, this.onDateSelected, this.getSelected})
-      : firstOfMonth = month,
+  _SKCalendarBody({DateTime? month, this.onDateSelected, this.getSelected})
+      : firstOfMonth = month!,
         startDate = month.weekday == 7
             ? month
             : DateTime(month.year, month.month, 1 - month.weekday);
@@ -567,11 +570,11 @@ class _SKCalendarBody extends StatefulWidget {
 }
 
 class _SKCalendarBodyState extends State<_SKCalendarBody> {
-  DateTime selectedDay;
+  late DateTime selectedDay;
 
   Widget build(BuildContext context) {
     final startDate = widget.startDate;
-    selectedDay = widget.getSelected();
+    selectedDay = widget.getSelected!();
 
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -623,7 +626,7 @@ class _SKCalendarBodyState extends State<_SKCalendarBody> {
               aspectRatio: 1,
               child: GestureDetector(
                 onTapUp: (details) {
-                  widget.onDateSelected(date);
+                  widget.onDateSelected!(date);
                   // selectedDay = date;
                 },
                 child: Container(
@@ -651,7 +654,7 @@ class _SKCalendarBodyState extends State<_SKCalendarBody> {
 class SKNavFadeUpRoute extends ModalRoute<void> {
   final WidgetBuilder builder;
 
-  SKNavFadeUpRoute({@required this.builder});
+  SKNavFadeUpRoute({required this.builder});
 
   @override
   Duration get transitionDuration => Duration(milliseconds: 500);
@@ -666,7 +669,7 @@ class SKNavFadeUpRoute extends ModalRoute<void> {
   Color get barrierColor => Colors.white;
 
   @override
-  String get barrierLabel => null;
+  String get barrierLabel => '';
 
   @override
   bool get maintainState => true;
@@ -689,7 +692,9 @@ class SKNavFadeUpRoute extends ModalRoute<void> {
     return FadeTransition(
       opacity: animation,
       child: CupertinoFullscreenDialogTransition(
-        animation: animation,
+        primaryRouteAnimation: animation,
+        secondaryRouteAnimation: secondaryAnimation,
+        linearTransition: false,
         child: child,
       ),
     );
@@ -700,7 +705,7 @@ class SKNavOverlayRoute extends ModalRoute<Object> {
   final WidgetBuilder builder;
   final bool isBarrierDismissible;
 
-  SKNavOverlayRoute({@required this.builder, this.isBarrierDismissible = true});
+  SKNavOverlayRoute({required this.builder, this.isBarrierDismissible = true});
 
   @override
   Duration get transitionDuration => Duration(milliseconds: 500);
@@ -715,7 +720,7 @@ class SKNavOverlayRoute extends ModalRoute<Object> {
   Color get barrierColor => Colors.black.withOpacity(0.2);
 
   @override
-  String get barrierLabel => null;
+  String get barrierLabel => '';
 
   @override
   bool get maintainState => true;
@@ -738,7 +743,9 @@ class SKNavOverlayRoute extends ModalRoute<Object> {
     return FadeTransition(
       opacity: animation,
       child: CupertinoFullscreenDialogTransition(
-        animation: animation,
+        primaryRouteAnimation: animation,
+        secondaryRouteAnimation: secondaryAnimation,
+        linearTransition: false,
         child: child,
       ),
     );
@@ -749,7 +756,7 @@ class SKCoverSheetNav extends ModalRoute<Object> {
   final WidgetBuilder builder;
   final bool isBarrierDismissible;
 
-  SKCoverSheetNav({@required this.builder, this.isBarrierDismissible = true});
+  SKCoverSheetNav({required this.builder, this.isBarrierDismissible = true});
 
   @override
   Duration get transitionDuration => Duration(milliseconds: 300);
@@ -764,7 +771,7 @@ class SKCoverSheetNav extends ModalRoute<Object> {
   Color get barrierColor => Colors.black.withOpacity(0.2);
 
   @override
-  String get barrierLabel => null;
+  String get barrierLabel => '';
 
   @override
   bool get maintainState => true;
@@ -800,7 +807,7 @@ class SKCoverSheetNav extends ModalRoute<Object> {
 class SKNoAnimationRoute extends ModalRoute<Object> {
   final WidgetBuilder builder;
 
-  SKNoAnimationRoute({@required this.builder});
+  SKNoAnimationRoute({required this.builder});
 
   @override
   Duration get transitionDuration => Duration(milliseconds: 0);
@@ -815,7 +822,7 @@ class SKNoAnimationRoute extends ModalRoute<Object> {
   Color get barrierColor => Colors.black;
 
   @override
-  String get barrierLabel => null;
+  String get barrierLabel => '';
 
   @override
   bool get maintainState => true;
@@ -851,13 +858,13 @@ class SKLoadingScreen extends ModalRoute<void> {
   Color get barrierColor => SKColors.text_light_gray.withOpacity(0.2);
 
   @override
-  String get barrierLabel => null;
+  String get barrierLabel => '';
 
   @override
   bool get maintainState => true;
 
   void fadeOut() {
-    navigator.pop();
+    navigator!.pop();
   }
 
   @override
@@ -887,17 +894,17 @@ class SKLoadingScreen extends ModalRoute<void> {
 
 class SKAlertDialog extends StatelessWidget {
   final String title;
-  final String subTitle;
+  final String? subTitle;
 
-  final String confirmText;
-  final String cancelText;
+  final String? confirmText;
+  final String? cancelText;
 
-  final Widget child;
+  final Widget? child;
 
-  final DynamicCallback getResults;
+  final DynamicCallback? getResults;
 
   SKAlertDialog({
-    @required this.title,
+    required this.title,
     this.subTitle,
     this.child,
     this.confirmText,
@@ -926,7 +933,7 @@ class SKAlertDialog extends StatelessWidget {
             Padding(
               padding: EdgeInsets.symmetric(vertical: 4, horizontal: 12),
               child: Text(
-                subTitle,
+                subTitle!,
                 textAlign: TextAlign.center,
                 style: TextStyle(
                     fontWeight: FontWeight.normal,
@@ -941,7 +948,7 @@ class SKAlertDialog extends StatelessWidget {
               height: 1,
               color: SKColors.border_gray,
             ),
-          if (child != null) child,
+          if (child != null) child!,
           Row(
             children: <Widget>[
               Expanded(
@@ -977,7 +984,7 @@ class SKAlertDialog extends StatelessWidget {
                 child: GestureDetector(
                   onTapUp: (details) => Navigator.pop(
                     context,
-                    getResults == null ? true : getResults(),
+                    getResults == null ? true : getResults!(),
                   ),
                   child: Container(
                     decoration: BoxDecoration(
@@ -1009,8 +1016,8 @@ class SKAlertDialog extends StatelessWidget {
 }
 
 class SKColorPicker extends StatefulWidget {
-  final ColorCallback callback;
-  final Widget child;
+  final ColorCallback? callback;
+  final Widget? child;
   final List<Color> colors = [
     Color(0xFFAE77BD),
     Color(0xFFE882AC),
@@ -1029,17 +1036,17 @@ class SKColorPicker extends StatefulWidget {
 }
 
 class _SKColorPickerState extends State<SKColorPicker> {
-  OverlayEntry _overlayEntry;
+  OverlayEntry? _overlayEntry;
 
   void _tappedColorPicker(TapUpDetails details) async {
     if (this._overlayEntry == null) {
       this._overlayEntry = _createOverlayEntry();
-      Overlay.of(context).insert(this._overlayEntry);
+      Overlay.of(context)!.insert(this._overlayEntry!);
     }
   }
 
   OverlayEntry _createOverlayEntry() {
-    RenderBox renderBox = context.findRenderObject();
+    RenderBox renderBox = context.findRenderObject() as RenderBox;
 
     final size = renderBox.size;
     final offset = renderBox.localToGlobal(Offset.zero);
@@ -1054,7 +1061,7 @@ class _SKColorPickerState extends State<SKColorPicker> {
     return OverlayEntry(
       builder: (context) => GestureDetector(
         onTapUp: (details) {
-          this._overlayEntry.remove();
+          this._overlayEntry!.remove();
           this._overlayEntry = null;
         },
         child: Stack(
@@ -1101,8 +1108,8 @@ class _SKColorPickerState extends State<SKColorPicker> {
 
   Widget _createColorCard(Color color) => GestureDetector(
         onTapUp: (details) {
-          widget.callback(color);
-          this._overlayEntry.remove();
+          widget.callback!(color);
+          this._overlayEntry!.remove();
           this._overlayEntry = null;
         },
         child: Container(
@@ -1161,18 +1168,18 @@ class _CustomArrowPainter extends CustomPainter {
 
 class SKPickerModal extends StatefulWidget {
   final String title;
-  final String subtitle;
+  final String? subtitle;
   final List<String> items;
   final IntCallback onSelect;
-  final Widget optionChild;
-  final VoidCallback onOptionChildTapped;
+  final Widget? optionChild;
+  final VoidCallback? onOptionChildTapped;
   final int startIndex;
 
   SKPickerModal({
-    @required this.title,
+    required this.title,
     this.subtitle,
-    @required this.items,
-    @required this.onSelect,
+    required this.items,
+    required this.onSelect,
     this.optionChild,
     this.onOptionChildTapped,
     this.startIndex = 0,
@@ -1184,7 +1191,7 @@ class SKPickerModal extends StatefulWidget {
 
 class _SKPickerModalState extends State<SKPickerModal> {
   int _selectedIndex = 0;
-  ScrollController scrollController;
+  late FixedExtentScrollController scrollController;
 
   @override
   void initState() {
@@ -1208,7 +1215,7 @@ class _SKPickerModalState extends State<SKPickerModal> {
         Padding(
           padding: EdgeInsets.fromLTRB(24, 4, 24, 0),
           child: Text(
-            widget.subtitle,
+            widget.subtitle!,
             textAlign: TextAlign.center,
             style: TextStyle(
               fontSize: 13,
@@ -1258,7 +1265,7 @@ class _SKPickerModalState extends State<SKPickerModal> {
                 behavior: HitTestBehavior.opaque,
                 onTapUp: (_) {
                   Navigator.pop(context);
-                  widget.onOptionChildTapped();
+                  widget.onOptionChildTapped!();
                 },
                 child: Padding(
                   padding: EdgeInsets.symmetric(vertical: 4),
@@ -1317,12 +1324,15 @@ class SKHeaderProfilePhoto extends StatelessWidget {
           color: Colors.white,
           shape: BoxShape.circle,
           boxShadow: [BoxShadow(blurRadius: 3, color: Color(0x19000000))],
-          image: DecorationImage(
-            fit: BoxFit.fill,
-            image: SKUser.current.avatarUrl == null
-                ? AssetImage(ImageNames.peopleImages.static_profile)
-                : NetworkImage(SKUser.current.avatarUrl),
-          ),
+          image: SKUser.current?.avatarUrl == null
+              ? DecorationImage(
+                  fit: BoxFit.fill,
+                  image: AssetImage(ImageNames.peopleImages.static_profile),
+                )
+              : DecorationImage(
+                  fit: BoxFit.fill,
+                  image: NetworkImage(SKUser.current?.avatarUrl ?? ""),
+                ),
         ),
         height: 30,
         width: 30,
@@ -1344,13 +1354,13 @@ enum SammiPersonality {
 enum SammiSide { left, right }
 
 class SammiSpeechBubble extends StatelessWidget {
-  final SammiPersonality sammiPersonality;
+  final SammiPersonality? sammiPersonality;
   final Widget speechBubbleContents;
   final SammiSide sammiSide;
 
   SammiSpeechBubble({
-    @required this.sammiPersonality,
-    @required this.speechBubbleContents,
+    required this.sammiPersonality,
+    required this.speechBubbleContents,
     this.sammiSide = SammiSide.left,
   });
 
@@ -1403,7 +1413,7 @@ class SammiSpeechBubble extends StatelessWidget {
               mainAxisSize: MainAxisSize.min,
               children: leftArrow
                   ? [
-                      image,
+                      image!,
                       Flexible(
                         flex: 1,
                         fit: FlexFit.tight,
@@ -1434,13 +1444,13 @@ class SammiSpeechBubble extends StatelessWidget {
                           ),
                         ),
                       ),
-                      image,
+                      image!,
                     ],
             ),
     );
   }
 
-  Image _sammiImageBuilder() {
+  Image? _sammiImageBuilder() {
     switch (sammiPersonality) {
       case SammiPersonality.cool:
         return Image.asset(ImageNames.sammiImages.cool);
@@ -1455,15 +1465,14 @@ class SammiSpeechBubble extends StatelessWidget {
       case SammiPersonality.school:
         return Image.asset(ImageNames.sammiImages.smile);
       case SammiPersonality.jobsCool:
-        return Image.asset(ImageNames.sammiJobsImages.swag);
+        return Image.asset(ImageNames.sammiImages.big_smile);
       case SammiPersonality.jobsOoo:
-        return Image.asset(ImageNames.sammiJobsImages.stare);
+        return Image.asset(ImageNames.sammiImages.big_smile);
       case SammiPersonality.jobsLargeSmile:
-        return Image.asset(ImageNames.sammiJobsImages.large_big_smile);
+        return Image.asset(ImageNames.sammiImages.big_smile);
       case SammiPersonality.jobsSmile:
-        return Image.asset(ImageNames.sammiJobsImages.smile);
+        return Image.asset(ImageNames.sammiImages.big_smile);
     }
-    return null;
   }
 }
 
@@ -1542,11 +1551,11 @@ class _SammiArrowPainter extends CustomPainter {
 enum ImpactGraphSize { small, large }
 
 class SKAssignmentImpactGraph extends StatelessWidget {
-  final double completion;
+  final double? completion;
   final ImpactGraphSize size;
   final Color color;
 
-  final ImpactLevel level;
+  final ImpactLevel? level;
 
   SKAssignmentImpactGraph(this.completion, this.color,
       {this.size = ImpactGraphSize.large})
@@ -1561,16 +1570,16 @@ class SKAssignmentImpactGraph extends StatelessWidget {
     String impactDesc;
 
     if (this.level != null) {
-      level = this.level;
+      level = this.level!;
       impactDesc = ["None", "Low", "Medium", "High"][level.index];
     } else {
       if ((completion ?? 0.0) == 0.0) {
         level = ImpactLevel.none;
         impactDesc = 'None';
-      } else if (completion < 0.05) {
+      } else if (completion! < 0.05) {
         level = ImpactLevel.low;
         impactDesc = 'Low';
-      } else if (completion < 0.15) {
+      } else if (completion! < 0.15) {
         level = ImpactLevel.medium;
         impactDesc = 'Medium';
       } else {
@@ -1677,25 +1686,30 @@ class _SKImpactGraphPainter extends CustomPainter {
 
 class SKHeaderCard extends StatelessWidget {
   final Widget leftHeaderItem;
-  final Widget rightHeaderItem;
+  final Widget? rightHeaderItem;
   final List<Widget> children;
-  final EdgeInsets margin;
-  final EdgeInsets padding;
+  final EdgeInsets? margin;
+  final EdgeInsets? padding;
+  final Color? headerColor;
+  final Color? backgroundColor;
+  final Color? borderColor;
 
-  SKHeaderCard({
-    @required this.leftHeaderItem,
-    this.rightHeaderItem,
-    @required this.children,
-    this.margin,
-    this.padding,
-  });
+  SKHeaderCard(
+      {required this.leftHeaderItem,
+      this.rightHeaderItem,
+      required this.children,
+      this.margin,
+      this.padding,
+      this.headerColor,
+      this.backgroundColor,
+      this.borderColor});
 
   @override
   Widget build(BuildContext context) => Container(
         decoration: BoxDecoration(
-          color: Colors.white,
+          color: this.backgroundColor ?? Colors.white,
           borderRadius: BorderRadius.circular(10),
-          border: Border.all(color: SKColors.border_gray),
+          border: Border.all(color: this.borderColor ?? SKColors.border_gray),
           boxShadow: UIAssets.boxShadow,
         ),
         margin: margin ?? EdgeInsets.all(16),
@@ -1706,7 +1720,7 @@ class SKHeaderCard extends StatelessWidget {
               alignment: Alignment.centerLeft,
               padding: EdgeInsets.fromLTRB(12, 12, 12, 8),
               decoration: BoxDecoration(
-                color: SKColors.selected_gray,
+                color: this.headerColor ?? SKColors.selected_gray,
                 borderRadius: BorderRadius.only(
                   topLeft: Radius.circular(10),
                   topRight: Radius.circular(10),
@@ -1716,7 +1730,7 @@ class SKHeaderCard extends StatelessWidget {
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
                   leftHeaderItem,
-                  if (rightHeaderItem != null) rightHeaderItem,
+                  if (rightHeaderItem != null) rightHeaderItem!,
                 ],
               ),
             ),
@@ -1734,7 +1748,7 @@ class SKHeaderCard extends StatelessWidget {
 }
 
 class ShakeAnimation extends StatefulWidget {
-  final Widget child;
+  final Widget? child;
 
   ShakeAnimation({this.child}) : super(key: UniqueKey());
 
@@ -1744,8 +1758,8 @@ class ShakeAnimation extends StatefulWidget {
 
 class _ShakeAnimationState extends State<ShakeAnimation>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation _animation;
+  late AnimationController _controller;
+  late Animation _animation;
 
   @override
   void initState() {
@@ -1807,8 +1821,8 @@ class GifWrapper extends StatefulWidget {
 }
 
 class _GifWrapperState extends State<GifWrapper> with TickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<int> _animation;
+  late AnimationController _controller;
+  late Animation<int> _animation;
 
   @override
   void initState() {
@@ -1831,7 +1845,7 @@ class _GifWrapperState extends State<GifWrapper> with TickerProviderStateMixin {
   Widget build(BuildContext context) {
     return AnimatedBuilder(
       animation: _animation,
-      builder: (BuildContext context, Widget child) {
+      builder: (BuildContext context, Widget? child) {
         String frame = _animation.value.toString().padLeft(3, '0');
         return Image.asset(
           'image_assets/${widget.baseGifName}/frame_${frame}_delay-0.04s.png',
@@ -1839,5 +1853,24 @@ class _GifWrapperState extends State<GifWrapper> with TickerProviderStateMixin {
         );
       },
     );
+  }
+}
+
+class CornerRadiusClipper extends CustomClipper<Path> {
+  final double radius;
+
+  CornerRadiusClipper(this.radius);
+
+  @override
+  Path getClip(Size size) {
+    final path = Path();
+    final rect = Rect.fromLTRB(0.0, 0.0, size.width, size.height);
+    path.addRRect(RRect.fromRectAndRadius(rect, Radius.circular(radius)));
+    return path;
+  }
+
+  @override
+  bool shouldReclip(CustomClipper<Path> oldClipper) {
+    return true;
   }
 }
