@@ -1,14 +1,15 @@
-import 'package:dropdown_banner/dropdown_banner.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:skoller/tools.dart';
-import 'package:skoller/screens/auth/phone_verification_view.dart';
-import 'package:url_launcher/url_launcher.dart';
 import 'package:dart_notification_center/dart_notification_center.dart';
+import 'package:dropdown_banner/dropdown_banner.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:skoller/screens/auth/phone_verification_view.dart';
+import 'package:skoller/tools.dart';
+import 'package:url_launcher/url_launcher.dart';
+
 import 'sign_in.dart';
 
 class SignUp extends StatefulWidget {
-  SignUp({Key key}) : super(key: key);
+  SignUp({Key? key}) : super(key: key);
 
   @override
   _SignUpState createState() => _SignUpState();
@@ -105,7 +106,7 @@ class _SignUpState extends State<SignUp> {
         }
       } else {
         setState(() => validState = false);
-        
+
         String message = [422, 401].contains(response.status)
             ? 'A user already exists with that phone number or email. Try logging in!'
             : 'Failed to create account. Try again later, or visit skoller.co';
@@ -119,7 +120,7 @@ class _SignUpState extends State<SignUp> {
     });
   }
 
-  double dist;
+  double? dist;
 
   @override
   Widget build(BuildContext context) => Scaffold(
@@ -134,7 +135,7 @@ class _SignUpState extends State<SignUp> {
                 onVerticalDragCancel: () => dist = null,
                 onVerticalDragEnd: (details) => dist = null,
                 onVerticalDragUpdate: (details) {
-                  if (dist != null && details.localPosition.dy - dist > 30) {
+                  if (dist != null && details.localPosition.dy - dist! > 30) {
                     [
                       firstNameFocus,
                       lastNameFocus,
@@ -215,8 +216,10 @@ class _SignUpState extends State<SignUp> {
                                     textCapitalization:
                                         TextCapitalization.words,
                                     textInputAction: TextInputAction.next,
-                                    onSubmitted: (_) =>
-                                        firstNameFocus.nextFocus(),
+                                    onSubmitted: (_) {
+                                      lastNameFocus.requestFocus();
+                                      // firstNameFocus.nextFocus();
+                                    },
                                     autofocus: true,
                                     onChanged: verifyState,
                                     focusNode: firstNameFocus,
@@ -261,8 +264,10 @@ class _SignUpState extends State<SignUp> {
                                     textCapitalization:
                                         TextCapitalization.words,
                                     textInputAction: TextInputAction.next,
-                                    onSubmitted: (_) =>
-                                        lastNameFocus.nextFocus(),
+                                    onSubmitted: (_) {
+                                      emailFocus.requestFocus();
+                                      // return lastNameFocus.nextFocus();
+                                    },
                                     onChanged: verifyState,
                                     focusNode: lastNameFocus,
                                   ),
@@ -309,7 +314,10 @@ class _SignUpState extends State<SignUp> {
                               decoration: BoxDecoration(border: null),
                               textCapitalization: TextCapitalization.none,
                               textInputAction: TextInputAction.next,
-                              onSubmitted: (_) => emailFocus.nextFocus(),
+                              onSubmitted: (_) {
+                                // return emailFocus.nextFocus();
+                                phoneFocus.requestFocus();
+                              },
                               onChanged: verifyState,
                               focusNode: emailFocus,
                             ),
@@ -379,7 +387,7 @@ class _SignUpState extends State<SignUp> {
                               TextSpan(
                                   text: 'User Agreement',
                                   style:
-                                      TextStyle(color: SKColors.skoller_blue))
+                                      TextStyle(color: SKColors.skoller_blue1))
                             ],
                           ),
                           textAlign: TextAlign.center,
@@ -426,7 +434,7 @@ class _SignUpState extends State<SignUp> {
                                   child: Text(
                                     ' Log In',
                                     style: TextStyle(
-                                        color: SKColors.skoller_blue,
+                                        color: SKColors.skoller_blue1,
                                         fontWeight: FontWeight.bold),
                                   ),
                                 ),
