@@ -1,13 +1,13 @@
-import 'package:dart_notification_center/dart_notification_center.dart';
-import 'package:dropdown_banner/dropdown_banner.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
-import './modals/weight_creation_modal.dart';
-import 'package:skoller/tools.dart';
-
 import 'dart:collection';
 
+import 'package:dart_notification_center/dart_notification_center.dart';
+import 'package:dropdown_banner/dropdown_banner.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:skoller/screens/main_app/classes/assignment_weight_view.dart';
+import 'package:skoller/tools.dart';
+
+import './modals/weight_creation_modal.dart';
 
 class _StateKeeper {
   bool isPoints = false;
@@ -28,16 +28,16 @@ class WeightExtractionView extends StatefulWidget {
 class _WeightExtractionState extends State<WeightExtractionView> {
   final pageController = PageController(initialPage: 0);
 
-  StudentClass studentClass;
+  StudentClass? studentClass;
 
   int currentView = 0;
-  _StateKeeper state;
+  late _StateKeeper state;
 
   @override
   void initState() {
     studentClass = StudentClass.currentClasses[widget.classId];
     state = _StateKeeper();
-    studentClass.acquireWeightLock();
+    studentClass!.acquireWeightLock();
     super.initState();
   }
 
@@ -85,7 +85,7 @@ class _WeightExtractionState extends State<WeightExtractionView> {
     );
 
     if (result is bool && result) {
-      studentClass.releaseDIYLock(isCompleted: false);
+      studentClass!.releaseDIYLock(isCompleted: false);
       Navigator.pop(context);
     }
   }
@@ -93,8 +93,8 @@ class _WeightExtractionState extends State<WeightExtractionView> {
   @override
   Widget build(BuildContext context) {
     return SKNavView(
-      title: studentClass.name,
-      titleColor: studentClass.getColor(),
+      title: studentClass!.name!,
+      titleColor: studentClass!.getColor(),
       callbackLeft: popUnsuccessfully,
       children: <Widget>[
         Padding(
@@ -201,7 +201,7 @@ class _SubviewOneState extends State<_SubviewOne> {
     if (!shouldContinue) return;
 
     final loader = SKLoadingScreen.fadeIn(context);
-    studentClass.createWeights(
+    studentClass!.createWeights(
       false,
       [
         {'name': 'All Assignments', 'value': 100}
@@ -408,7 +408,7 @@ class _SubviewTwoState extends State<_SubviewTwo> {
               height: 36,
               decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(5),
-                  color: SKColors.skoller_blue,
+                  color: SKColors.skoller_blue1,
                   boxShadow: UIAssets.boxShadow),
               child: Text(
                 'Next',
@@ -472,7 +472,7 @@ class _SubviewThreeState extends State<_SubviewThree> {
     });
   }
 
-  Future<bool> showWeightMaker(
+  Future<dynamic> showWeightMaker(
     String nameStr,
     String valueStr,
     bool isCreate,
@@ -507,7 +507,7 @@ class _SubviewThreeState extends State<_SubviewThree> {
     final state = widget.subviewParent.state;
     final studentClass = widget.subviewParent.studentClass;
 
-    studentClass.createWeights(state.isPoints, state.weights).then((success) {
+    studentClass!.createWeights(state.isPoints, state.weights).then((success) {
       //After creating
       if (success) {
         return studentClass.releaseDIYLock(isCompleted: true);
