@@ -97,7 +97,7 @@ class ModModal extends StatelessWidget {
           mainAxisSize: MainAxisSize.min,
           children: <Widget>[
             Text(
-              parentClass.name,
+              parentClass.name!,
               textAlign: TextAlign.center,
               style: TextStyle(
                 color: parentClass.getColor(),
@@ -120,7 +120,7 @@ class ModModal extends StatelessWidget {
               ),
             ),
             Text(
-              assignment.name,
+              assignment.name??'',
               style: TextStyle(
                 color: parentClass.getColor(),
                 fontWeight: FontWeight.w700,
@@ -155,10 +155,10 @@ class ModModal extends StatelessWidget {
 
   Widget typeChange() {
     if (mod.modType == ModType.due) {
-      final oldDate = mod.parentAssignment.due;
+      final oldDate = mod.parentAssignment?.due;
       final newDate = mod.data as DateTime;
 
-      final oldOrdinal = getOrdinal(oldDate);
+      final oldOrdinal = getOrdinal(oldDate!);
       final newOrdinal = getOrdinal(newDate);
 
       return Row(
@@ -224,9 +224,9 @@ class ModModal extends StatelessWidget {
                       fontSize: 13),
                 ),
                 Text(
-                  mod.parentAssignment.weight_id == null
+                  mod.parentAssignment?.weight_id == null
                       ? 'Not graded'
-                      : mod.parentAssignment.weightObject.name,
+                      : mod.parentAssignment!.weightObject.name,
                   style: TextStyle(color: SKColors.light_gray),
                 ),
               ],
